// src/modules/disciplines-annotation/dto/discipline-annotation.dto.ts
export interface DisciplineAnnotationDTO {
  electoral: boolean;
  level: string;
  amount: number;
  course: number;
  language: string;
  cathedra: string;
  requirements: string;
  courseProgram: string;
  reasonsToStudy: string;
  studyResult: string;
  usages: string;
  materials: string;
  formOfConducting: string;
  semesterControl: string;
  minStudents: number;
  maxStudents: number;
  maxContracts: number;
}